Move common code to activity_stream package
