+++ datahub/activity_stream/tests/utils.py
@@ -1,0 +1,8 @@
+from rest_framework.reverse import reverse
+
+
+def get_url(name):
+    """
+    Returns the URL for test server given the endpoint name.
+    """
+    return 'http://testserver' + reverse(name)
